Fix the transaction details tests to render deeply
